# guile: fix Apple Silicon JIT compilation