@fxcoudert 2022-11-17
Please provide context so we understand the reason.

- Why is the patch needed? The original report, by me, was for guile 3.0.4 two years ago, and resulted in a crash during build. Right now, guile 3.0.8 builds fine on Apple Silicon (both Big Sur, Monterey and Ventura).
- Has the patch been committed to guile? Or reviewed and approved, at least?