@aconchillo 2022-11-17
Hi!

> Please provide context so we understand the reason.
> 
> * Why is the patch needed? The original report, by me, was for guile 3.0.4 two years ago, and resulted in a crash during build. Right now, guile 3.0.8 builds fine on Apple Silicon (both Big Sur, Monterey and Ventura).

The current formula disables JIT compiler on Apple Silicon because it crashed as you mentioned. The reason of the crash is because on Apple Silicon you need to follow this when generating JIT code:

https://developer.apple.com/documentation/apple-silicon/porting-just-in-time-compilers-to-apple-silicon

Which Guile was not doing and therefore when JIT code was being generated (during build) it crashed. This still happens in 3.0.8 and upstream.

The provided patch fixes this issue and enables JIT compilation on Apple Silicon which makes Guile MUCH faster. For example, see run times at end of this example:

```
(define-syntax-rule (do-times n exp)
  (let lp ((count n))
    (let ((count (1- count)))
      exp
      (unless (zero? count) (lp count)))))

(define (loop-to-1e4)
  (let lp ((i 0))
    (when (< i [#e1e4](https://fosstodon.org/tags/e1e4))
      (lp (1+ i)))))

,t (do-times 100000 (loop-to-1e4))

Without JIT:
;; 13.081576s real time, 13.078927s run time.  0.000000s spent in GC.

With JIT:
;; 0.742761s real time, 0.741818s run time.  0.000000s spent in GC.
```

> * Has the patch been committed to guile? Or reviewed and approved, at least?

No, it has not. It's been sent to guile-devel@gnu.org:

https://lists.gnu.org/archive/html/guile-devel/2022-11/msg00017.html

Unfortunately patches sent to Guile might take a long time to be reviewed. I believe the patch is safe and follows Apple's recommendations that's why I sent this PR and I think it's better to have a faster Guile :smiley: .