@aconchillo 2022-11-17
> That's what Debian thought about a faster OpenSSL: https://www.debian.org/security/2008/dsa-1571
> 
> I think it should at least be reviewed upstream before we accept it.

Sounds good. I'll keep you updated.

 Thanks!