@fxcoudert 2022-12-10
Let's close this for now, and we'll consider a pull request when the patch is committed upstream.